requirements.txt
pandas
pyserial
requests
pyubx2

